chore: refactor configuration handling
- Remove the import of `github.com/blackhorseya/ekko/internal/pkg/config`
- Replace `newConfigCmd` function call with `config.NewConfigCmd`
- Delete the `adapter/cli/app/config.go` file
- Create a new file `adapter/cli/app/config/config.go`
