+++ adapter/cli/app/cmd.go
@@ -1,8 +1,8 @@
 package app
 
 import (
+	"github.com/blackhorseya/ekko/adapter/cli/app/config"
 	"github.com/blackhorseya/ekko/adapter/cli/app/version"
-	"github.com/blackhorseya/ekko/internal/pkg/config"
 	"github.com/blackhorseya/ekko/pkg/adapters"
 	"github.com/spf13/cobra"
 )
@@ -12,14 +12,14 @@ type cmd struct {
 }
 
 // NewCmd is used to create a new cmd instance
-func NewCmd(config *config.Config) adapters.CLI {
+func NewCmd() adapters.CLI {
 	rootCmd := &cobra.Command{
 		Short:        "ekko is a tool for todo list management",
 		SilenceUsage: true,
 	}
 
 	rootCmd.AddCommand(version.NewVersionCmd())
-	rootCmd.AddCommand(newConfigCmd(config))
+	rootCmd.AddCommand(config.NewConfigCmd())
 
 	return &cmd{
 		rootCmd: rootCmd,
